changed the select_all method and refactored
